auth-service/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
python-dotenv==1.0.0
prometheus-client==0.17.1
bcrypt==4.0.1  # Для хеширования паролей